+++ MiniPubSub/PubSub/MessageMediator.swift
@@ -1,6 +1,6 @@
 //
 //  MessageMediator.swift
-//  iOSCore
+//  MiniPubSub
 //
 //  Created by sangmin park on 1/19/24.
 //