+++ MiniPubSub/PubSub/MessageMediatorImpl.swift
@@ -1,6 +1,6 @@
 //
 //  MessageMediatorImpl.swift
-//  iOSCore
+//  MiniPubSub
 //
 //  Created by sangmin park on 1/19/24.
 //