+++ MiniPubSub/PubSub/Node.swift
@@ -1,0 +1,13 @@
+//
+//  Node.swift
+//  MiniPubSub
+//
+//  Created by sangmin park on 1/19/24.
+//
+
+import Foundation
+
+public protocol Node{
+    var id: Int { get }
+}
+