+++ MiniPubSub/iOSCore.h
@@ -7,11 +7,11 @@
 
 #import <Foundation/Foundation.h>
 
-//! Project version number for iOSCore.
-FOUNDATION_EXPORT double iOSCoreVersionNumber;
+//! Project version number for MiniPubSub.
+FOUNDATION_EXPORT double MiniPubSubVersionNumber;
 
 //! Project version string for iOSCore.
-FOUNDATION_EXPORT const unsigned char iOSCoreVersionString[];
+FOUNDATION_EXPORT const unsigned char MiniPubSubVersionString[];
 
 // In this header, you should import all the public headers of your framework using statements like #import <iOSCore/PublicHeader.h>